+++ kompendium-annotations/Module.md
@@ -1,0 +1,9 @@
+# Module kompendium-annotations
+
+This module houses all annotations that Kompendium uses to provide key metadata when performing reflective analysis. 
+
+It is separated from core predominantly to allow for potential future integrations with [Kotlin Symbol Processing](https://github.com/google/ksp)
+
+# Package io.bkbn.kompendium.annotations
+
+Contains all annotations used by Kompendium
