+++ kompendium-annotations/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+  id("kotlin-library-conventions")
+}
+
+tasks.dokkaHtmlPartial.configure {
+  dokkaSourceSets {
+    configureEach {
+      includes.from("Module.md")
+    }
+  }
+}
