+++ kompendium-annotations/src/main/kotlin/io/bkbn/kompendium/annotations/KompendiumField.kt
@@ -1,0 +1,9 @@
+package io.bkbn.kompendium.annotations
+
+/**
+ * Annotation used to perform field level overrides.
+ * @param name Indicates that a field name override is desired.  Often used for camel case to snake case conversions.
+ */
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.PROPERTY)
+annotation class KompendiumField(val name: String)
