+++ kompendium-annotations/src/main/kotlin/io/bkbn/kompendium/annotations/ParamType.kt
@@ -1,0 +1,11 @@
+package io.bkbn.kompendium.annotations
+
+/**
+ * The allowed parameter types as specified by the OpenAPI specification
+ */
+enum class ParamType {
+  COOKIE,
+  HEADER,
+  PATH,
+  QUERY
+}
