+++ kompendium-auth/src/test/kotlin/io/bkbn/kompendium/auth/util/TestModels.kt
@@ -1,8 +1,8 @@
 package io.bkbn.kompendium.auth.util
 
-import io.bkbn.kompendium.core.annotations.KompendiumField
-import io.bkbn.kompendium.core.annotations.KompendiumParam
-import io.bkbn.kompendium.core.annotations.ParamType
+import io.bkbn.kompendium.annotations.KompendiumField
+import io.bkbn.kompendium.annotations.KompendiumParam
+import io.bkbn.kompendium.annotations.ParamType
 
 data class TestParams(
     @KompendiumParam(ParamType.PATH) val a: String,
