+++ kompendium-auth/src/test/resources/notarized_multiple_jwt_authenticated_get.json
@@ -54,10 +54,12 @@
   "components" : {
     "securitySchemes" : {
       "jwt1" : {
+        "bearerFormat" : "JWT",
         "type" : "http",
         "scheme" : "bearer"
       },
       "jwt2" : {
+        "bearerFormat" : "JWT",
         "type" : "http",
         "scheme" : "bearer"
       }
