--- kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/annotations/KompendiumField.kt
@@ -1,5 +1,0 @@
-package io.bkbn.kompendium.core.annotations
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.PROPERTY)
-annotation class KompendiumField(val name: String)
