--- kompendium-core/src/main/kotlin/io/bkbn/kompendium/core/annotations/UndeclaredField.kt
@@ -1,8 +1,0 @@
-package io.bkbn.kompendium.core.annotations
-
-import kotlin.reflect.KClass
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.CLASS)
-@Repeatable
-annotation class UndeclaredField(val field: String, val clazz: KClass<*>)
