+++ kompendium-playground/src/main/kotlin/io/bkbn/kompendium/playground/LocationsSpike.kt
@@ -1,8 +1,8 @@
 package io.bkbn.kompendium.playground
 
 import io.bkbn.kompendium.core.Kompendium
-import io.bkbn.kompendium.core.annotations.KompendiumParam
-import io.bkbn.kompendium.core.annotations.ParamType
+import io.bkbn.kompendium.annotations.KompendiumParam
+import io.bkbn.kompendium.annotations.ParamType
 import io.bkbn.kompendium.core.metadata.MethodInfo
 import io.bkbn.kompendium.core.metadata.ResponseInfo
 import io.bkbn.kompendium.core.routes.openApi
